requirements.txt
opencv-python==4.5.5.64
numpy==1.21.5
torch==1.10.0
pandas==1.3.5
ultralytics==8.0.20
Pillow==9.0.1
transformers==4.18.0
scikit-image==0.19.2
filterpy==1.4.5
scikit-learn==1.0.2